feat: 更新 .gitignore 和新增文档
- 更新 .gitignore，添加对 dist-ssr、*.local、*.pid、*.pid.lock 和上传目录的忽略。
- 新增社区小程序的全链路电商与用户体系升级大纲及需求文档。
- 新增社区数据库表字段中文说明文档和全表全字段清单。
- 新增 O2O 集市后端联调接口清单文档。
- 新增惠民卡前后端对接与后端实现需求文档。
- 新增惠民卡与线上社区 API 对接说明文档。
- 新增惠民卡京东/拼多多联盟数据库与同步说明文档。
- 新增技工后台与服务商后台的 Web 管理端需求说明文档。

+++ backend/src/models/MarketShopCategory.js
@@ -5,7 +5,7 @@ module.exports = (sequelize, DataTypes) => {
     static associate(models) {}
   }
   MarketShopCategory.init({
-    shop_id: { type: DataTypes.BIGINT, allowNull: false },
+    shop_id: { type: DataTypes.INTEGER, allowNull: false },
     category_key: { type: DataTypes.STRING(50), allowNull: false },
     category_name: { type: DataTypes.STRING(50), allowNull: false },
     sort_order: { type: DataTypes.INTEGER, allowNull: false, defaultValue: 0 },
